test: 패키지 정리
