+++ src/test/java/kitchenpos/orders/eatinorders/application/EatInOrderServiceTest.java
@@ -1,16 +1,12 @@
-package kitchenpos.orders.application.takoutorders.application;
+package kitchenpos.orders.eatinorders.application;
 
 import kitchenpos.menus.application.InMemoryMenuRepository;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.order.domain.*;
-import kitchenpos.order.eatinorders.domain.OrderTableRepository;
-import kitchenpos.order.takeoutorders.application.TakeOutOrderService;
-import kitchenpos.order.takeoutorders.domain.TakeOutOrderAcceptService;
-import kitchenpos.order.takeoutorders.domain.TakeOutOrderCompleteService;
-import kitchenpos.order.takeoutorders.domain.TakeOutOrderCreateService;
-import kitchenpos.order.takeoutorders.domain.TakeOutOrderServeService;
-import kitchenpos.orders.application.InMemoryOrderRepository;
-import kitchenpos.orders.application.InMemoryOrderTableRepository;
+import kitchenpos.order.eatinorders.application.EatInOrderService;
+import kitchenpos.order.eatinorders.domain.*;
+import kitchenpos.orders.InMemoryOrderRepository;
+import kitchenpos.orders.eatinorders.infra.InMemoryOrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -28,21 +24,21 @@ import static kitchenpos.Fixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-class TakeOutOrderServiceTest {
+class EatInOrderServiceTest {
     private OrderRepository orderRepository;
     private MenuRepository menuRepository;
     private OrderTableRepository orderTableRepository;
     private OrderLineItemsValidService orderLineItemsValidService;
 
 
-    // TAKE_OUT order service
-    private TakeOutOrderService takeOutOrderService;
-    private TakeOutOrderCreateService takeOutOrderCreateService;
-    private TakeOutOrderAcceptService takeOutOrderAcceptService;
-    private TakeOutOrderServeService takeOutOrderServeService;
-    private TakeOutOrderCompleteService takeOutOrderCompleteService;
-
+    // EAT_IN order service
+    private EatInOrderService eatInOrderService;
+    private EatInOrderCreateService eatInOrderCreateService;
+    private EatInOrderAcceptService eatInOrderAcceptService;
+    private EatInOrderServeService eatInOrderServeService;
+    private EatInOrderCompleteService eatInOrderCompleteService;
 
     private static List<Arguments> orderLineItems() {
         return Arrays.asList(
@@ -62,38 +58,42 @@ class TakeOutOrderServiceTest {
         menuRepository = new InMemoryMenuRepository();
         orderTableRepository = new InMemoryOrderTableRepository();
         orderLineItemsValidService = new OrderLineItemsValidService(menuRepository);
-        takeOutOrderCreateService = new TakeOutOrderCreateService(orderRepository, orderLineItemsValidService);
-        takeOutOrderAcceptService = new TakeOutOrderAcceptService(orderRepository);
-        takeOutOrderServeService = new TakeOutOrderServeService(orderRepository);
-        takeOutOrderCompleteService = new TakeOutOrderCompleteService(orderRepository);
-        takeOutOrderService = new TakeOutOrderService(orderRepository, takeOutOrderCreateService, takeOutOrderAcceptService, takeOutOrderServeService, takeOutOrderCompleteService);
+
+        eatInOrderCreateService = new EatInOrderCreateService(orderTableRepository, orderLineItemsValidService);
+        eatInOrderAcceptService = new EatInOrderAcceptService(orderRepository);
+        eatInOrderServeService = new EatInOrderServeService(orderRepository);
+        eatInOrderCompleteService = new EatInOrderCompleteService(orderRepository, new OrderTableClearService(orderTableRepository, orderRepository));
+        eatInOrderService = new EatInOrderService(orderRepository, eatInOrderCreateService, eatInOrderAcceptService, eatInOrderServeService, eatInOrderCompleteService);
 
     }
 
-    @DisplayName("1개 이상의 등록된 메뉴로 포장 주문을 등록할 수 있다.")
+
+    @DisplayName("1개 이상의 등록된 메뉴로 매장 주문을 등록할 수 있다.")
     @Test
-    void createTakeoutOrder() {
+    void createEatInOrder() {
         final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(OrderType.TAKEOUT, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        final Order actual = takeOutOrderService.create(expected);
+        final UUID orderTableId = orderTableRepository.save(orderTable(true, 4)).getId();
+        final Order expected = createOrderRequest(OrderType.EAT_IN, orderTableId, createOrderLineItemRequest(menuId, 19_000L, 3L));
+        final Order actual = eatInOrderService.create(expected);
         assertThat(actual).isNotNull();
         assertAll(
                 () -> assertThat(actual.getId()).isNotNull(),
                 () -> assertThat(actual.getType()).isEqualTo(expected.getType()),
                 () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.WAITING),
                 () -> assertThat(actual.getOrderDateTime()).isNotNull(),
-                () -> assertThat(actual.getOrderLineItems().getOrderLineItems()).hasSize(1)
+                () -> assertThat(actual.getOrderLineItems().getOrderLineItems()).hasSize(1),
+                () -> assertThat(actual.getOrderTable().getId()).isEqualTo(expected.getOrderTableId())
         );
     }
-
 
     @DisplayName("주문 유형이 올바르지 않으면 등록할 수 없다.")
     @NullSource
     @ParameterizedTest
     void create(final OrderType type) {
         final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(type, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        assertThatThrownBy(() -> takeOutOrderService.create(expected))
+        final UUID orderTableId = orderTableRepository.save(orderTable(true, 4)).getId();
+        final Order expected = createOrderRequest(type, orderTableId, createOrderLineItemRequest(menuId, 19_000L, 3L));
+        assertThatThrownBy(() -> eatInOrderService.create(expected))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -101,30 +101,43 @@ class TakeOutOrderServiceTest {
     @MethodSource("orderLineItems")
     @ParameterizedTest
     void create(final List<OrderLineItem> orderLineItems) {
+        final UUID orderTableId = orderTableRepository.save(orderTable(false, 0)).getId();
         assertThatThrownBy(() -> {
-            final Order expected = createOrderRequest(OrderType.TAKEOUT, orderLineItems);
-            takeOutOrderService.create(expected);
+            final Order expected = createOrderRequest(OrderType.EAT_IN, orderTableId, orderLineItems);
+            eatInOrderService.create(expected);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
-    @DisplayName("매장 주문을 제외한 주문의 경우 주문 항목의 수량은 0 이상이어야 한다.")
+    @DisplayName("매장 주문은 주문 항목의 수량이 0 미만일 수 있다.")
     @ValueSource(longs = -1L)
     @ParameterizedTest
-    void createWithoutEatInOrder(final long quantity) {
+    void createEatInOrder(final long quantity) {
         final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final UUID orderTableId = orderTableRepository.save(orderTable(true, 4)).getId();
+        final Order expected = createOrderRequest(
+                OrderType.EAT_IN, orderTableId, createOrderLineItemRequest(menuId, 19_000L, quantity)
+        );
+        assertDoesNotThrow(() -> eatInOrderService.create(expected));
+    }
 
-        assertThatThrownBy(() -> {
-            final Order expected = createOrderRequest(OrderType.TAKEOUT, createOrderLineItemRequest(menuId, 19_000L, quantity));
-            takeOutOrderService.create(expected);
-        }).isInstanceOf(IllegalArgumentException.class);
+    @DisplayName("빈 테이블에는 매장 주문을 등록할 수 없다.")
+    @Test
+    void createEmptyTableEatInOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
+        final UUID orderTableId = orderTableRepository.save(orderTable(false, 0)).getId();
+        final Order expected = createOrderRequest(
+                OrderType.EAT_IN, orderTableId, createOrderLineItemRequest(menuId, 19_000L, 3L)
+        );
+        assertThatThrownBy(() -> eatInOrderService.create(expected))
+                .isInstanceOf(IllegalStateException.class);
     }
 
     @DisplayName("숨겨진 메뉴는 주문할 수 없다.")
     @Test
     void createNotDisplayedMenuOrder() {
         final UUID menuId = menuRepository.save(menu(19_000L, false, menuProduct())).getId();
-        final Order expected = createOrderRequest(OrderType.TAKEOUT, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        assertThatThrownBy(() -> takeOutOrderService.create(expected))
+        final Order expected = createOrderRequest(OrderType.EAT_IN, createOrderLineItemRequest(menuId, 19_000L, 3L));
+        assertThatThrownBy(() -> eatInOrderService.create(expected))
                 .isInstanceOf(IllegalStateException.class);
     }
 
@@ -132,8 +145,8 @@ class TakeOutOrderServiceTest {
     @Test
     void createNotMatchedMenuPriceOrder() {
         final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final Order expected = createOrderRequest(OrderType.TAKEOUT, createOrderLineItemRequest(menuId, 16_000L, 3L));
-        assertThatThrownBy(() -> takeOutOrderService.create(expected))
+        final Order expected = createOrderRequest(OrderType.EAT_IN, createOrderLineItemRequest(menuId, 16_000L, 3L));
+        assertThatThrownBy(() -> eatInOrderService.create(expected))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -141,7 +154,7 @@ class TakeOutOrderServiceTest {
     @Test
     void accept() {
         final UUID orderId = orderRepository.save(order(OrderStatus.WAITING, orderTable(true, 4))).getId();
-        final Order actual = takeOutOrderService.accept(orderId);
+        final Order actual = eatInOrderService.accept(orderId);
         assertThat(actual.getStatus()).isEqualTo(OrderStatus.ACCEPTED);
     }
 
@@ -150,7 +163,7 @@ class TakeOutOrderServiceTest {
     @ParameterizedTest
     void accept(final OrderStatus status) {
         final UUID orderId = orderRepository.save(order(status, orderTable(true, 4))).getId();
-        assertThatThrownBy(() -> takeOutOrderService.accept(orderId))
+        assertThatThrownBy(() -> eatInOrderService.accept(orderId))
                 .isInstanceOf(IllegalStateException.class);
     }
 
@@ -159,7 +172,7 @@ class TakeOutOrderServiceTest {
     @Test
     void serve() {
         final UUID orderId = orderRepository.save(order(OrderStatus.ACCEPTED)).getId();
-        final Order actual = takeOutOrderService.serve(orderId);
+        final Order actual = eatInOrderService.serve(orderId);
         assertThat(actual.getStatus()).isEqualTo(OrderStatus.SERVED);
     }
 
@@ -168,7 +181,7 @@ class TakeOutOrderServiceTest {
     @ParameterizedTest
     void serve(final OrderStatus status) {
         final UUID orderId = orderRepository.save(order(status)).getId();
-        assertThatThrownBy(() -> takeOutOrderService.serve(orderId))
+        assertThatThrownBy(() -> eatInOrderService.serve(orderId))
                 .isInstanceOf(IllegalStateException.class);
     }
 
@@ -178,18 +191,44 @@ class TakeOutOrderServiceTest {
     @ParameterizedTest
     void completeTakeoutAndEatInOrder(final OrderStatus status) {
         final UUID orderId = orderRepository.save(order(status)).getId();
-        assertThatThrownBy(() -> takeOutOrderService.complete(orderId))
+        assertThatThrownBy(() -> eatInOrderService.complete(orderId))
                 .isInstanceOf(IllegalStateException.class);
     }
 
+    @DisplayName("주문 테이블의 모든 매장 주문이 완료되면 빈 테이블로 설정한다.")
+    @Test
+    void completeEatInOrder() {
+        final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
+        final Order expected = orderRepository.save(order(OrderStatus.SERVED, orderTable));
+        final Order actual = eatInOrderService.complete(expected.getId());
+        assertAll(
+                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.COMPLETED),
+                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().isOccupied()).isFalse(),
+                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().getNumberOfGuests()).isEqualTo(0)
+        );
+    }
+
+    @DisplayName("완료되지 않은 매장 주문이 있는 주문 테이블은 빈 테이블로 설정하지 않는다.")
+    @Test
+    void completeNotTable() {
+        final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
+        orderRepository.save(order(OrderStatus.ACCEPTED, orderTable));
+        final Order expected = orderRepository.save(order(OrderStatus.SERVED, orderTable));
+        final Order actual = eatInOrderService.complete(expected.getId());
+        assertAll(
+                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.COMPLETED),
+                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().isOccupied()).isTrue(),
+                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().getNumberOfGuests()).isEqualTo(4)
+        );
+    }
 
     @DisplayName("주문의 목록을 조회할 수 있다.")
     @Test
     void findAll() {
         final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
         orderRepository.save(order(OrderStatus.SERVED, orderTable));
-        orderRepository.save(order(OrderStatus.DELIVERED, "서울시 송파구 위례성대로 2"));
-        final List<Order> actual = takeOutOrderService.findAll();
+        orderRepository.save(order(OrderStatus.COMPLETED, "서울시 송파구 위례성대로 2"));
+        final List<Order> actual = eatInOrderService.findAll();
         assertThat(actual).hasSize(2);
     }
 
@@ -216,5 +255,13 @@ class TakeOutOrderServiceTest {
             final OrderLineItem... orderLineItems
     ) {
         return new Order(UUID.randomUUID(), type, OrderStatus.WAITING, LocalDateTime.now(), new OrderLineItems(Arrays.asList(orderLineItems)), null, null, orderTableId);
+    }
+
+    private Order createOrderRequest(
+            final OrderType type,
+            final UUID orderTableId,
+            final List<OrderLineItem> orderLineItems
+    ) {
+        return new Order(UUID.randomUUID(), type, OrderStatus.WAITING, LocalDateTime.now(), new OrderLineItems(orderLineItems), null, null, orderTableId);
     }
 }
